8_Building_Your_First_GenAI_Application_with_AWS_Data_Foundations/8.3_Building_Your_Life_Insurance_Agent/lambda/initialize-db/requirements.txt
# lambda/db-initializer/requirements.txt
psycopg2-binary
boto3
urllib3